+++ classes/StancerApiPayment.php
@@ -31,6 +31,9 @@ class StancerApiPayment extends ObjectModel
     /** @var int Order id */
     public $id_order;
 
+    /** @var bool Is a live mode object? */
+    public $live_mode;
+
     /** @var string Currency */
     public $currency;
 
@@ -82,6 +85,11 @@ class StancerApiPayment extends ObjectModel
             'id_order' => [
                 'type' => self::TYPE_INT,
                 'validate' => 'isUnsignedId',
+            ],
+            'live_mode' => [
+                'required' => true,
+                'type' => self::TYPE_BOOL,
+                'validate' => 'isBool',
             ],
             'currency' => [
                 'required' => true,
@@ -175,6 +183,7 @@ class StancerApiPayment extends ObjectModel
         $query->from(static::$definition['table']);
         $query->where('`id_cart` = ' . (int) $cart->id);
         $query->where('`currency` = "' . pSQL($currency->iso_code) . '"');
+        $query->where('`status` = "pending"');
 
         $row = Db::getInstance()->getRow($query);
         if (!$row) {
@@ -236,7 +245,7 @@ class StancerApiPayment extends ObjectModel
     }
 
     /**
-     * Save Stancer api payment
+     * Save Stancer API payment
      *
      * @param bool $null_values
      * @param bool $auto_date
@@ -248,8 +257,31 @@ class StancerApiPayment extends ObjectModel
      */
     public function save($null_values = false, $auto_date = true): bool
     {
+        $config = new StancerApiConfig();
+
+        $this->live_mode = $config->isLiveMode();
+
         if ($this->api) {
-            $this->payment_id = $this->api->getId();
+            $this->payment_id = $this->api->id;
+
+            if ($this->api->card) {
+                $this->card_id = $this->api->card->id;
+            }
+
+            if ($this->api->customer) {
+                $this->customer_id = $this->api->customer->id;
+            }
+
+            $status = $this->api->status;
+
+            if (!$status && $this->api->auth) {
+                if (in_array($this->api->auth->status, ['declined', 'expired', 'failed', 'unavailable'], true)) {
+                    // We can not mark the payment failed in the API
+                    $status = Stancer\Payment\Status::FAILED;
+                }
+            }
+
+            $this->status = $status ?? 'pending';
         }
 
         return parent::save($null_values, $auto_date);
@@ -266,24 +298,24 @@ class StancerApiPayment extends ObjectModel
     public static function saveFrom(Stancer\Payment $apiPayment, Cart $cart): StancerApiPayment
     {
         $payment = static::findByApiPayment($apiPayment);
+
         if (!$payment) {
             $payment = new static();
         }
 
-        $card = $apiPayment->getCard();
-        $creation = $apiPayment->getCreationDate();
-        $customer = $apiPayment->getCustomer();
+        $card = $apiPayment->card;
+        $creation = $apiPayment->creationDate;
+        $customer = $apiPayment->customer;
 
-        $payment->payment_id = $apiPayment->getId();
-        $payment->currency = $apiPayment->getCurrency();
-        $payment->amount = $apiPayment->getAmount();
-        $payment->status = $apiPayment->getStatus() ?: 'pending';
-        $payment->card_id = $card ? $card->getId() : null;
-        $payment->created = $creation ? $creation->format('Y-m-d H:i:s') : null;
-        $payment->customer_id = $customer ? $customer->getId() : null;
-        $payment->id_cart = $cart->id;
-
+        $payment->amount = $apiPayment->amount;
         $payment->api = $apiPayment;
+        $payment->card_id = $card ? $card->id : null;
+        $payment->created = $creation ? $creation->format('Y-m-d H:i:s') : null;
+        $payment->currency = $apiPayment->currency;
+        $payment->customer_id = $customer ? $customer->id : null;
+        $payment->id_cart = $cart->id;
+        $payment->payment_id = $apiPayment->id;
+
         $payment->save();
 
         return $payment;